Improve argument checking
Clarified and expanded checks for which combinations of input are in use, as proposed by @jalfje.

We improve the argument check section, and collect all the errors and warnings on an `ArgCheck` object from ArgumentCheck package (which is now added to the DESCRIPTION file). The function will now cast more descriptive errors including all the missing
or incorrect arguments. The function will also cast a message indicating the options used for computing ETo. A new argument (`verbose`) is added to the function which allows for suppressing
this message (default set to TRUE, so it will show the message).