Correct computation of solar declination angle in thornthwaite()
Solving an issue by @nikhil003. Now Delta and tan(Delta) in thornthwaite() are computed instead of being hard-coded. This results in slightly different results than before, as the tanDelta vector used before was a bad approximation. Resulting PET values are nearly identical, but the code is more clear and correct now.

Also, adding a comparison between PET / ETo methods in the man page.